ingestor-sql/requirements.txt
pandas
mysql-connector-python
sqlalchemy
boto3
python-dotenv
